chore: move sign or send into the wallet.ts file
